+++ packages/contracts/contracts/Interfaces/IActivePool.sol
@@ -9,12 +9,12 @@ interface IActivePool is IPool {
     // --- Events ---
     event BorrowerOperationsAddressChanged(address _newBorrowerOperationsAddress);
     event TroveManagerAddressChanged(address _newTroveManagerAddress);
-    event ActivePoolTHUSDDebtUpdated(uint _THUSDDebt);
-    event ActivePoolCollateralBalanceUpdated(uint _collateral);
+    event ActivePoolTHUSDDebtUpdated(uint256 _THUSDDebt);
+    event ActivePoolCollateralBalanceUpdated(uint256 _collateral);
     event CollateralAddressChanged(address _newCollateralAddress);
     event CollSurplusPoolAddressChanged(address _newCollSurplusPoolAddress);
 
     // --- Functions ---
-    function sendCollateral(address _account, uint _amount) external;
+    function sendCollateral(address _account, uint256 _amount) external;
     function updateCollateralBalance(uint256 _amount) external;
 }
