+++ packages/contracts/contracts/Interfaces/IBorrowerOperations.sol
@@ -19,9 +19,9 @@ interface IBorrowerOperations {
     event PCVAddressChanged(address _pcvAddress);
     event CollateralAddressChanged(address _newCollateralAddress);
 
-    event TroveCreated(address indexed _borrower, uint arrayIndex);
-    event TroveUpdated(address indexed _borrower, uint _debt, uint _coll, uint stake, uint8 operation);
-    event THUSDBorrowingFeePaid(address indexed _borrower, uint _THUSDFee);
+    event TroveCreated(address indexed _borrower, uint256 arrayIndex);
+    event TroveUpdated(address indexed _borrower, uint256 _debt, uint256 _coll, uint256 stake, uint8 operation);
+    event THUSDBorrowingFeePaid(address indexed _borrower, uint256 _THUSDFee);
 
     // --- Functions ---
 
@@ -40,23 +40,23 @@ interface IBorrowerOperations {
 
     ) external;
 
-    function openTrove(uint _maxFee, uint _THUSDAmount, uint _assetAmount, address _upperHint, address _lowerHint) external payable;
+    function openTrove(uint256 _maxFee, uint256 _THUSDAmount, uint256 _assetAmount, address _upperHint, address _lowerHint) external payable;
 
-    function addColl(uint _assetAmount, address _upperHint, address _lowerHint) external payable;
+    function addColl(uint256 _assetAmount, address _upperHint, address _lowerHint) external payable;
 
-    function moveCollateralGainToTrove(address _user, uint _assetAmount, address _upperHint, address _lowerHint) external payable;
+    function moveCollateralGainToTrove(address _user, uint256 _assetAmount, address _upperHint, address _lowerHint) external payable;
 
-    function withdrawColl(uint _amount, address _upperHint, address _lowerHint) external;
+    function withdrawColl(uint256 _amount, address _upperHint, address _lowerHint) external;
 
-    function withdrawTHUSD(uint _maxFee, uint _amount, address _upperHint, address _lowerHint) external;
+    function withdrawTHUSD(uint256 _maxFee, uint256 _amount, address _upperHint, address _lowerHint) external;
 
-    function repayTHUSD(uint _amount, address _upperHint, address _lowerHint) external;
+    function repayTHUSD(uint256 _amount, address _upperHint, address _lowerHint) external;
 
     function closeTrove() external;
 
-    function adjustTrove(uint _maxFee, uint _collWithdrawal, uint _debtChange, bool isDebtIncrease, uint _assetAmount, address _upperHint, address _lowerHint) external payable;
+    function adjustTrove(uint256 _maxFee, uint256 _collWithdrawal, uint256 _debtChange, bool isDebtIncrease, uint256 _assetAmount, address _upperHint, address _lowerHint) external payable;
 
     function claimCollateral() external;
 
-    function getCompositeDebt(uint _debt) external pure returns (uint);
+    function getCompositeDebt(uint256 _debt) external pure returns (uint);
 }
