+++ packages/contracts/contracts/Interfaces/ITHUSDToken.sol
@@ -13,7 +13,7 @@ interface ITHUSDToken is IERC20, IERC2612 {
     event StabilityPoolAddressChanged(address _newStabilityPoolAddress);
     event BorrowerOperationsAddressChanged(address _newBorrowerOperationsAddress);
 
-    event THUSDTokenBalanceUpdated(address _user, uint _amount);
+    event THUSDTokenBalanceUpdated(address _user, uint256 _amount);
 
     // --- Functions ---
