+++ packages/contracts/contracts/THUSDToken.sol
@@ -240,8 +240,8 @@ contract THUSDToken is Ownable, CheckContract, ITHUSDToken {
     (
         address owner,
         address spender,
-        uint amount,
-        uint deadline,
+        uint256 amount,
+        uint256 deadline,
         uint8 v,
         bytes32 r,
         bytes32 s
